Backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.4
boto3==1.42.14
botocore==1.42.14
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
jmespath==1.0.1
kombu==5.6.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
